Unversion zones controller so it can be used with Versioncake for testing purposes
