+++ api/app/views/spree/api/zones/index.rabl
@@ -1,6 +1,6 @@
 object false
 child(@zones => :zones) do
-  extends 'spree/api/v1/zones/show'
+  extends 'spree/api/zones/show'
 end
 node(:count) { @zones.count }
 node(:current_page) { params[:page] || 1 }
