+++ api/app/views/spree/api/zones/show.rabl
